+++ apps/sasquatch/src/main/java/com/microsoft/appcenter/sasquatch/activities/SettingsActivity.java
@@ -43,6 +43,7 @@ import com.microsoft.appcenter.sasquatch.eventfilter.EventFilter;
 import com.microsoft.appcenter.utils.PrefStorageConstants;
 
 import java.io.File;
+import java.lang.reflect.Method;
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
 import java.util.Date;
@@ -151,6 +152,40 @@ public class SettingsActivity extends AppCompatActivity {
                     onDatabaseFileChanged(dbFile);
                 }
             };
+            initCheckBoxSetting(R.string.appcenter_network_requests_allowed_key, R.string.appcenter_network_requests_allowed, R.string.appcenter_network_requests_disallowed, new HasEnabled() {
+
+                @Override
+                public void setEnabled(boolean enabled) {
+                    try {
+
+                        /*
+                         * TODO replace the next line with 'AppCenter.setNetworkRequestsAllowed(enabled);'
+                         * when updating the demo during release process.
+                         */
+                        Method setNetworkRequestsAllowedMethod = AppCenter.class.getMethod("setNetworkRequestsAllowed", boolean.class);
+                        setNetworkRequestsAllowedMethod.invoke(null, enabled);
+                    } catch (Exception e) {
+                        Log.d(LOG_TAG, "No setNetworkRequestsAllowed API in this build");
+                    }
+                }
+
+                @Override
+                public boolean isEnabled() {
+                    try {
+
+                        /*
+                         * TODO replace the next line with 'AppCenter.isNetworkRequestsAllowed();'
+                         * when updating the demo during release process.
+                         */
+                        Method isNetworkRequestsAllowedMethod = AppCenter.class.getMethod("isNetworkRequestsAllowed");
+                        Object value = isNetworkRequestsAllowedMethod.invoke(null);
+                        return (boolean)value;
+                    } catch (Exception e) {
+                        Log.d(LOG_TAG, "No isNetworkRequestsAllowed API in this build");
+                        return true;
+                    }
+                }
+            });
 
             /* Analytics. */
             initCheckBoxSetting(R.string.appcenter_analytics_state_key, R.string.appcenter_analytics_state_summary_enabled, R.string.appcenter_analytics_state_summary_disabled, new HasEnabled() {