+++ sdk/appcenter/src/main/java/com/microsoft/appcenter/channel/Channel.java
@@ -135,6 +135,13 @@ public interface Channel {
     void shutdown();
 
     /**
+     * Set network request allowed. If true check pending logs, suspend sending logs otherwise.
+     *
+     * @param isAllowed True if network request allowed, false otherwise.
+     */
+    void setNetworkRequests(boolean isAllowed);
+
+    /**
      * Channel global listener specification.
      */
     interface Listener {